pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bioMed"
version = "0.1.0"
description = "OpenEnv environment for PET bioremediation planning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core]>=0.2.1",
  "uvicorn>=0.24.0",
]

[project.scripts]
server = "server.app:main"
biomed-server = "server.app:main"

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "ruff",
  "mypy",
]

[tool.setuptools]
py-modules = ["client"]

[tool.setuptools.packages.find]
include = ["models", "server", "server.*", "training", "training.*"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "unit: fast deterministic unit tests",
  "integration: in-process subsystem integration tests",
  "api: HTTP API contract tests",
  "e2e: slower end-to-end benchmark and deployment tests",
  "slow: slow-running tests that should not block the default fast lane",
]
